+++ packages/backend/src/const.ts
@@ -3,6 +3,8 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
+import {RolePolicies} from "@/core/RoleService.js";
+
 export const MAX_NOTE_TEXT_LENGTH = 3000;
 
 export const USER_ONLINE_THRESHOLD = 1000 * 60 * 10; // 10min
@@ -26,7 +28,7 @@ export const DB_MAX_IMAGE_COMMENT_LENGTH = 512;
 
 // ブラウザで直接表示することを許可するファイルの種類のリスト
 // ここに含まれないものは application/octet-stream としてレスポンスされる
-// SVGはXSSを生むので許可しない
+// SVG は XSS を生むので許可しない
 export const FILE_TYPE_BROWSERSAFE = [
 	// Images
 	'image/png',
@@ -74,3 +76,29 @@ https://github.com/sindresorhus/file-type/blob/main/supported.js
 https://github.com/sindresorhus/file-type/blob/main/core.js
 https://developer.mozilla.org/en-US/docs/Web/Media/Formats/Containers
 */
+export const DEFAULT_POLICIES: RolePolicies = {
+	gtlAvailable: true,
+	ltlAvailable: true,
+	canPublicNote: true,
+	canInvite: false,
+	inviteLimit: 0,
+	inviteLimitCycle: 60 * 24 * 7,
+	inviteExpirationTime: 0,
+	canManageCustomEmojis: false,
+	canManageAvatarDecorations: false,
+	canSearchNotes: false,
+	canUseTranslator: true,
+	canHideAds: false,
+	driveCapacityMb: 100,
+	alwaysMarkNsfw: false,
+	pinLimit: 5,
+	antennaLimit: 5,
+	wordMuteLimit: 200,
+	webhookLimit: 3,
+	clipLimit: 10,
+	noteEachClipsLimit: 200,
+	userListLimit: 10,
+	userEachUserListsLimit: 50,
+	rateLimitFactor: 1,
+	avatarDecorationLimit: 1,
+} as const;
